redesign stack run view and fix broken viewport on logs scroll
