+++ assets/src/components/stacks/stack/run/Sidecar.tsx
@@ -1,74 +1,24 @@
 import { ReactNode } from 'react'
-import {
-  Button,
-  ReloadIcon,
-  Sidecar,
-  SidecarItem,
-} from '@pluralsh/design-system'
+import { Sidecar, SidecarItem } from '@pluralsh/design-system'
 
 import { useTheme } from 'styled-components'
 import moment from 'moment'
-import { useNavigate, useParams } from 'react-router-dom'
-import { getStackRunsAbsPath } from 'routes/stacksRoutesConsts'
-import { GqlError } from 'components/utils/Alert'
 
-import {
-  StackRun,
-  StackStatus,
-  useApproveStackRunMutation,
-  useRestartStackRunMutation,
-  useUpdateStackRunMutation,
-} from '../../../../generated/graphql'
+import { StackRun } from '../../../../generated/graphql'
 import { ResponsiveLayoutSidecarContainer } from '../../../utils/layout/ResponsiveLayoutSidecarContainer'
 import UserInfo from '../../../utils/UserInfo'
 import { ClusterProviderIcon } from '../../../utils/Provider'
 import { StackRunStatusChip } from '../../common/StackRunStatusChip'
+import StackApprovalChip from '../../common/StackApprovalChip'
 
 interface StackRunSidecarProps {
   stackRun: StackRun
-  refetch?: Nullable<() => void>
 }
-
-const TERMINAL_STATES = [
-  StackStatus.Successful,
-  StackStatus.Cancelled,
-  StackStatus.Failed,
-]
 
 export default function StackRunSidecar({
   stackRun,
-  refetch,
 }: StackRunSidecarProps): ReactNode {
-  const { stackId } = useParams()
   const theme = useTheme()
-  const navigate = useNavigate()
-
-  const [mutation, { loading, error }] = useApproveStackRunMutation({
-    variables: { id: stackRun.id },
-    onCompleted: () => refetch?.(),
-  })
-
-  const [restart, { loading: restartLoading, error: restartError }] =
-    useRestartStackRunMutation({
-      variables: { id: stackRun.id },
-      onCompleted: ({ restartStackRun }) =>
-        navigate(getStackRunsAbsPath(stackId, restartStackRun?.id)),
-    })
-
-  const [cancel, { loading: cancelLoading }] = useUpdateStackRunMutation({
-    variables: {
-      id: stackRun.id,
-      attributes: {
-        status: StackStatus.Cancelled,
-      },
-    },
-    onCompleted: () => refetch?.(),
-  })
-
-  const terminal = TERMINAL_STATES.includes(stackRun.status)
-
-  if (error) return <GqlError error={error} />
-  if (restartError) return <GqlError error={restartError} />
 
   return (
     <ResponsiveLayoutSidecarContainer
@@ -76,34 +26,6 @@ export default function StackRunSidecar({
       flexDirection="column"
       gap="small"
     >
-      {stackRun.status === StackStatus.PendingApproval &&
-        !stackRun.approvedAt && (
-          <Button
-            onClick={mutation}
-            loading={loading}
-          >
-            Approve Run
-          </Button>
-        )}
-      {terminal && (
-        <Button
-          secondary
-          onClick={restart}
-          loading={restartLoading}
-          startIcon={<ReloadIcon />}
-        >
-          Restart Run
-        </Button>
-      )}
-      {!terminal && (
-        <Button
-          secondary
-          onClick={cancel}
-          loading={cancelLoading}
-        >
-          Cancel
-        </Button>
-      )}
       <Sidecar css={{ overflowX: 'auto' }}>
         <SidecarItem heading="Status">
           <StackRunStatusChip
@@ -111,8 +33,11 @@ export default function StackRunSidecar({
             size="small"
           />
         </SidecarItem>
-        <SidecarItem heading="Needs approval">
-          {stackRun.approval ? 'Required' : 'Not required'}
+        <SidecarItem heading="Approval">
+          <StackApprovalChip
+            approval={stackRun.approval ?? false}
+            size="small"
+          />
         </SidecarItem>
         {stackRun.approvedAt && (
           <SidecarItem heading="Approved at">
@@ -124,9 +49,6 @@ export default function StackRunSidecar({
             <UserInfo user={stackRun.approver} />
           </SidecarItem>
         )}
-        {stackRun.message && (
-          <SidecarItem heading="Commit message">{stackRun.message}</SidecarItem>
-        )}
         <SidecarItem heading="Cluster">
           <div css={{ display: 'flex', gap: theme.spacing.xsmall }}>
             <ClusterProviderIcon
@@ -135,9 +57,6 @@ export default function StackRunSidecar({
             />
             {stackRun?.cluster?.name}
           </div>
-        </SidecarItem>
-        <SidecarItem heading="Repository">
-          {stackRun.repository?.url}
         </SidecarItem>
         <SidecarItem heading="Created at">
           {moment(stackRun.insertedAt).format('lll')}
