+++ assets/src/components/stacks/stack/run/progress/Progress.tsx
@@ -35,7 +35,6 @@ export default function StackRunProgress(): ReactNode {
   return (
     <ScrollablePage
       scrollable={false}
-      heading="Progress"
       noPadding
     >
       <Card
